[ddraw.ini] Change of default settings, replace Russian comments, add Keycodes
MouseSensitivity: change to 100 because why wouldn't it start at 100%?

ReloadWeaponKey: changed to null because implementing a shortcut key
like this is not a good idea. "W" is a single-stroke shortcut so if a
user presses it accidentally and the item gets used up, they'll have no
idea what the hell happened unless they specifically already read about
this feature and its shortcut

PlayIdlAnimOnReload: set to 1 because this is a cool feature :)

ProcessorIdle: set to 1 because excessive CPU usage is a problem with
this game

AutoReloadWeapon: set to 0 because this is not something that would be
universally desired. Also, further similar reasoning to ReloadWeaponKey
above.

ReloadReserve: set to -1 vanilla; see also AutoReloadWeapon and
ReloadWeaponKey.

ColorLOS: set to 1 (default) and notated which is default. See also
AutoReloadWeapon, ReloadWeaponKey, ReloadReserve.

RemoveFriendlyFoe: set to 1 because that Perk was such a waste; default
in Fallout 2 as it should be. Nice job on this fix :)